Fish palettes - groups colors into their palettes The name of the palette is set here

Fish palettes - groups colors into their palettes The name of the palette is set here

Function that extracts hex codes form the fishpals_colors vector by name

get_fishpals_cols(...)

get_fishpals_cols(...)

Arguments

...
Quoted color name or vector of names.

Value

The 6-digit hexes attached to the name

Examples

get_fishpals_cols()

#>            juvsteel green            juvsteel taupe             juvsteel grey 
#>                 "#A3C595"                 "#BA8F78"                 "#5D5360" 
#>     juvsteel jungle green       juvsteel tuscan red     adultsteel dark liver 
#>                 "#1c240d"                 "#784039"                 "#504C41" 
#>      adultsteel red ochre           adultsteel gray       adultsteel feldgrau 
#>                 "#E2704D"                 "#B8BABD"                 "#4B5859" 
#> adultsteel spanish bistre               chinook red             chinook olive 
#>                 "#707239"                 "#BA1D57"                 "#8DAD6E" 
#>         chinook dark gray              chinook blue              chinook pink 
#>                 "#273535"                 "#0C4FA0"                 "#EFB0A1" 
#>             sunfish green               sunfish tan            sunfish yellow 
#>                 "#4EC4B8"                 "#8C7A4B"                 "#F1B24A" 
#>              sunfish gray         sunfish dark gray                  CFS blue 
#>                 "#636271"                 "#2D3040"                 "#1372AA" 
#>             CFS dark gray             CFS gridlines                   CFS tan 
#>                 "#3B4856"                 "#9FADBD"                 "#8B662E" 
#>             CFS light tan                  CFS teal           genidaqs yellow 
#>                 "#C3985E"                 "#1CD2A3"                 "#FAFF6B" 
#>             genidaqs blue            genidaqs brown            genidaqs green 
#>                 "#2597C4"                 "#3A3737"                 "#6CFC95" 
#>             genidaqs pink        surgewrasse orange          surgewrasse teal 
#>                 "#CA5B97"                 "#E24D4D"                 "#1C5B67" 
#>         surgewrasse green          surgewrasse blue        surgewrasse purple 
#>                 "#B6DD18"                 "#12DAF6"                 "#966480" 

get_fishpals_cols("chinook red")

#> chinook red 
#>   "#BA1D57"